split reporting checks into own file
